requirements.txt
altair==5.5.0
appscript==1.4.0
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cssselect2==0.8.0
defusedxml==0.7.1
docx2pdf==0.1.8
docxtpl==0.20.2
fonttools==4.60.1
fpdf2==2.8.5
gitdb==4.0.12
GitPython==3.1.45
google-generativeai==0.8.5
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pydyf==0.11.0
pypandoc==1.16.2
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
rpds-py==0.29.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
tinycss2==1.5.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
flask==2.3.3
flask-cors==4.0.0
torch==2.4.1
transformers==4.41.2
accelerate==0.30.1
sentencepiece==0.2.0
PyPDF2==3.0.1